+++ src/Runtime/DdevRuntimeHost.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Gaambo\DeployerWordpress\Runtime;
-
-/** @deprecated Use \Gaambo\DeployerUtils\Runtime\DdevRuntimeHost. */
-class DdevRuntimeHost extends \Gaambo\DeployerUtils\Runtime\DdevRuntimeHost
-{
-}