+++ src/Runtime/Runtime.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Gaambo\DeployerWordpress\Runtime;
-
-/** @deprecated Use \Gaambo\DeployerUtils\Runtime\Runtime. */
-class Runtime extends \Gaambo\DeployerUtils\Runtime\Runtime
-{
-}