+++ src/Runtime/RuntimeHost.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Gaambo\DeployerWordpress\Runtime;
-
-/** @deprecated Use \Gaambo\DeployerUtils\Runtime\RuntimeHost. */
-abstract class RuntimeHost extends \Gaambo\DeployerUtils\Runtime\RuntimeHost
-{
-}